backend/tafuta-api/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-cors-headers==3.12.0
django-extensions==3.1.5
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.3
imutils==0.5.4
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.3
openapi-codec==1.3.2
opencv-python==4.5.5.62
pandas==1.4.2
Pillow==9.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
